fixed reducers
