+++ src/components/Checkout/Checkout.js
@@ -1,25 +1,26 @@
-import React from 'react'
+import React from "react";
 // import { Link } from 'react-router-dom';
-import './Checkout.css'
-import { CartState } from '../../context/Context';
-
+import "./Checkout.css";
+import { CartState } from "../../context/Context";
 
 const Checkout = () => {
-
   const {
-    state: { cart }
+    state: { cart },
   } = CartState();
 
   // console.log(cart)
 
   const totalQty = cart.reduce((acc, curr) => acc + curr.qty, 0);
-const cartWithIndividualPrice = cart.map(item => {
-  return {
-    ...item,
-    price: item.price * item.qty
-  }
-});
-const totalPrice = cartWithIndividualPrice.reduce((acc, curr) => acc + curr.price, 0);
+  const cartWithIndividualPrice = cart.map((item) => {
+    return {
+      ...item,
+      price: item.price * item.qty,
+    };
+  });
+  const totalPrice = cartWithIndividualPrice.reduce(
+    (acc, curr) => acc + curr.price,
+    0
+  );
 
   return (
     <div className="checkout-page">
@@ -29,20 +30,24 @@ const totalPrice = cartWithIndividualPrice.reduce((acc, curr) => acc + curr.pric
         </div>
         <div class="address-details">
           <h2>Name</h2>
-          <p>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Incidunt facere a placeat officiis illum excepturi dolor sunt veniam maiores voluptates.</p>
+          <p>
+            Lorem ipsum dolor sit amet consectetur, adipisicing elit. Incidunt
+            facere a placeat officiis illum excepturi dolor sunt veniam maiores
+            voluptates.
+          </p>
         </div>
         <button>Add New Adress</button>
       </div>
       <div class="right-side-details">
-          <h1>Details</h1>
-          <div class="price-details">
+        <h1>Details</h1>
+        <div class="price-details">
           <p>Total Items in Your Cart : {totalQty}</p>
           <p>Total Price of cart : {totalPrice}</p>
-          </div>
-          <button>Pay Now</button>
+        </div>
+        <button>Pay Now</button>
       </div>
     </div>
-  )
-}
+  );
+};
 
-export default Checkout
+export default Checkout;
