+++ src/reducers/productReducer.js
@@ -1,0 +1,22 @@
+export const productReducer = (state, action) => {
+  switch (action.type) {
+    case "SORT_BY_PRICE":
+      return { ...state, sort: action.payload };
+    case "FILTER_BY_STOCK":
+      return { ...state, byStock: !state.byStock };
+    case "FILTER_BY_DELIVERY":
+      return { ...state, byFastDelivery: !state.byFastDelivery };
+    case "FILTER_BY_RATING":
+      return { ...state, byRating: action.payload };
+    case "FILTER_BY_SEARCH":
+      return { ...state, searchQuery: action.payload };
+    case "CLEAR_FILTERS":
+      return { byStock: false, byFastDelivery: false, byRating: 0 };
+    case "FILTER_BY_GENDER":
+      return { ...state, byGender: action.payload };
+    case "FILTER_BY_SIZE":
+      return { ...state, bySize: action.payload };
+    default:
+      return state;
+  }
+};
